Componentize new portfolio details and allocations
